version 1.0.12-beta
1. added singular extension
2. moved malus to before lazy move gen
3. saved check scores
4. removed tt flag in main negamax

+++ src/engine/param.h
@@ -61,12 +61,12 @@ constexpr bool IS_VALID(int16_t value)
     return value != VALUE_NONE;
 }
 
-constexpr int16_t MATED_IN(int16_t ply)
+constexpr int16_t MATED_IN(int32_t ply)
 {
     return -INF + ply;
 }
 
-constexpr int16_t MATE_IN(int16_t ply)
+constexpr int16_t MATE_IN(int32_t ply)
 {
     return INF - ply;
 }
